+++ src/index.js
@@ -1,6 +1,12 @@
-
-import ReactDOM from "react-dom"
 import React from 'react'
-import Kennel from "./components/Kennel"
+import ReactDOM from 'react-dom'
+import { BrowserRouter as Router } from "react-router-dom"
+import Kennel from './components/Kennel'
 
-ReactDOM.render(<Kennel />, document.querySelector("#root"));
+import './index.css'
+
+ReactDOM.render(
+    <Router>
+        <Kennel />
+    </Router>
+    , document.getElementById('root'))
